+++ yash-syntax/src/alias.rs
@@ -19,7 +19,7 @@
 //! This module provides data structures for defining aliases in the shell
 //! execution environment.
 
-use crate::source::Location;
+use crate::source::Span;
 use std::borrow::Borrow;
 use std::collections::HashSet;
 use std::hash::Hash;
@@ -35,9 +35,9 @@ pub struct Alias {
     pub replacement: String,
     /// Whether this alias is a global alias or not.
     pub global: bool,
-    /// Location of the word in the simple command that invoked the alias built-in to define this
+    /// Position of the word in the simple command that invoked the alias built-in to define this
     /// alias.
-    pub origin: Location,
+    pub origin: Span,
 }
 
 /// Wrapper of [`Alias`] for inserting into a hash set.
@@ -49,7 +49,7 @@ pub struct Alias {
 /// ```
 /// let mut entries = std::collections::HashSet::new();
 /// let name = "foo";
-/// let origin = yash_syntax::source::Location::dummy("");
+/// let origin = yash_syntax::source::Span::dummy("");
 /// let old = yash_syntax::alias::HashEntry::new(
 ///     name.to_string(), "old".to_string(), false, origin.clone());
 /// let new = yash_syntax::alias::HashEntry::new(
@@ -64,7 +64,7 @@ pub struct HashEntry(pub Rc<Alias>);
 
 impl HashEntry {
     /// Convenience method for creating a new alias definition as `HashEntry`
-    pub fn new(name: String, replacement: String, global: bool, origin: Location) -> HashEntry {
+    pub fn new(name: String, replacement: String, global: bool, origin: Span) -> HashEntry {
         HashEntry(Rc::new(Alias {
             name,
             replacement,